Reorganize functions related to editor
Remove redundant functions, and move functions that primarily deal with content
to their own file.
